requirements.txt
torch
torchvision
scikit-learn
numpy
pillow
matplotlib
PyQt6